Refactored code a bit more
