+++ .env.example
@@ -2,8 +2,3 @@ ETH_NODE_URL=https://mainnet.infura.io/v3/{API_KEY}
 
 PUBLIC_ADDRESS={PUBLIC_ADDRESS}
 PRIVATE_KEY={PRIVATE_KEY}
-
-# Gnosis Safe
-NODE_ENV="production"
-REACT_APP_NETWORK="production"
-REACT_APP_ENV="production"
